Backend/server.py
# 这是一个简单的后端示例，基于 FastAPI 和 Google Gemini
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import google.generativeai as genai
import os

# 初始化
app = FastAPI()

# 允许跨域 (让你的 React 前端能访问)
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_methods=["*"],
    allow_headers=["*"],
)

# 配置你的 API Key
GOOGLE_API_KEY = "你的_GOOGLE_API_KEY"
genai.configure(api_key=GOOGLE_API_KEY)

@app.get("/analyze/{video_id}")
async def analyze_video(video_id: str):
    try:
        # 1. 这里通常需要先下载 YouTube 视频 (使用 yt-dlp 库)
        # 为了演示简单，我们假设你已经有了视频文件，或者让 AI 分析视频的文本/关键帧
        # 实际上 Gemini 支持直接处理 YouTube 链接的能力正在逐步开放，
        # 现阶段最稳妥的做法是：后端下载视频 -> 上传给 Gemini -> 提问。
        
        # 模拟：构建一个提示词 (Prompt)
        prompt = """
        作为一个专业的视频分析师，请分析这个视频（假设已上传）：
        1. 提取脚本结构（时间戳+内容）。
        2. 分析动效风格（评分1-10，并解释）。
        3. 提取视觉配色风格。
        请以 JSON 格式返回。
        """

        # 2. 调用 Gemini API
        model = genai.GenerativeModel('gemini-1.5-pro-latest')
        # 注意：实际代码中这里需要上传视频文件对象
        # response = model.generate_content([video_file, prompt])
        
        # 3. 这里是模拟返回的 AI 真实数据
        return {
            "status": "success",
            "data": {
                "motion_score": 8.5,
                "motion_analysis": "检测到大量快速剪辑和动态排版，属于高燃风格。",
                "script_structure": [
                    {"time": "0:00", "label": "Hook: 提出痛点"},
                    {"time": "0:30", "label": "Body: 解决方案演示"}
                ]
            }
        }

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# 启动命令: uvicorn server:app --reload