requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
youtube-transcript-api==0.6.1
google-generativeai==0.3.2
requests==2.31.0
